+++ SilKit/source/capi/CapiLogger.cpp
@@ -25,4 +25,18 @@ SilKit_ReturnCode SilKit_Logger_Log(SilKit_Logger* self, SilKit_LoggingLevel lev
     CAPI_LEAVE
 }
 
+
+SilKit_ReturnCode SilKit_Logger_GetLogLevel(SilKit_Logger* self, SilKit_LoggingLevel* outLevel) {
+    ASSERT_VALID_POINTER_PARAMETER(self);
+    ASSERT_VALID_POINTER_PARAMETER(outLevel);
+    CAPI_ENTER
+    {
+        auto logger = reinterpret_cast<SilKit::Services::Logging::ILogger*>(self);
+        auto enumLevel = static_cast<SilKit_LoggingLevel>(logger->GetLogLevel());
+        *outLevel = enumLevel;
+        return SilKit_ReturnCode_SUCCESS;
+    }
+    CAPI_LEAVE
+}
+
 }