+++ SilKit/source/services/logging/Logger.cpp
@@ -201,9 +201,9 @@ void Logger::LogReceivedMsg(const LogMsg& msg)
     }
 }
 
-bool Logger::ShouldLog(Level level) const
+Level Logger::GetLogLevel() const
 {
-    return _logger->should_log(to_spdlog(level));
+    return from_spdlog(_logger->level());
 }
 
 } // namespace Logging